Low: crmd: Move to common function.
